Upgrade postgres_ext and make sure no new tests break
