+++ lib/postgres_ext/active_record/5.0/relation/predicate_builder/array_handler.rb
@@ -1,38 +1,22 @@
 require 'active_record/relation/predicate_builder'
 require 'active_record/relation/predicate_builder/array_handler'
 
-require 'active_support/concern'
-
-module ActiveRecord
-  class PredicateBuilder
-    module ArrayHandlerPatch
-      extend ActiveSupport::Concern
-
-      included do
-        def call_with_feature(attribute, value)
-          column = case attribute.try(:relation)
-                   when Arel::Nodes::TableAlias, NilClass
-                   else
-                     cache = ActiveRecord::Base.connection.schema_cache
-                     if cache.data_source_exists? attribute.relation.name
-                       cache.columns(attribute.relation.name).detect{ |col| col.name.to_s == attribute.name.to_s }
-                     end
-                   end
-          if column && column.respond_to?(:array) && column.array
-            attribute.eq(value)
-          else
-            call_without_feature(attribute, value)
-          end
-        end
-
-        alias_method_chain(:call, :feature)
-      end
-
-      module ClassMethods
-
-      end
+module ArrayHandlerPatch
+  def call(attribute, value)
+    column = case attribute.try(:relation)
+             when Arel::Nodes::TableAlias, NilClass
+             else
+               cache = ActiveRecord::Base.connection.schema_cache
+               if cache.data_source_exists? attribute.relation.name
+                 cache.columns(attribute.relation.name).detect{ |col| col.name.to_s == attribute.name.to_s }
+               end
+             end
+    if column && column.respond_to?(:array) && column.array
+      attribute.eq(value)
+    else
+      super(attribute, value)
     end
   end
 end
 
-ActiveRecord::PredicateBuilder::ArrayHandler.send(:include, ActiveRecord::PredicateBuilder::ArrayHandlerPatch)
+ActiveRecord::PredicateBuilder::ArrayHandler.send(:prepend, ArrayHandlerPatch)
