+++ lib/postgres_ext/active_record/relation/query_methods/with_chain.rb
@@ -1,0 +1,18 @@
+module ActiveRecord
+  module QueryMethods
+    # WithChain objects act as placeholder for queries in which #with does not have any parameter.
+    # In this case, #with must be chained with #recursive to return a new relation.
+    class WithChain
+      def initialize(scope)
+        @scope = scope
+      end
+
+      # Returns a new relation expressing WITH RECURSIVE
+      def recursive(*args)
+        @scope.with_values += args
+        @scope.recursive_value = true
+        @scope
+      end
+    end
+  end
+end
